Corrigé de la fonction set_cursor qui ne dépassait pas 255 à cause d'une variable de type uint8_t au lieu de uint16_t

+++ Programmes_LPC/StationDeGestion_TP_Train/src/Ecran.c
@@ -168,7 +168,7 @@ void Write_pixel(uint8_t red,uint8_t green,uint8_t blue)
 	Send_color(blue);
 }
 
-void Set_cursor(uint16_t x,uint8_t y)
+void Set_cursor(uint16_t x,uint16_t y)
 {
 	Index_out(0x20);
 	Parameter_out(x);
