+++ Programmes_LPC/StationDeGestion_TP_Train/src/Ecran.h
@@ -26,7 +26,7 @@ void Parameter_out(uint16_t param);
 void Send_color(uint8_t color);
 void Init_display();
 void Write_pixel(uint8_t red,uint8_t green,uint8_t blue);
-void Set_cursor(uint16_t x,uint8_t y);
+void Set_cursor(uint16_t x,uint16_t y);
 void Create_partial_screen(uint16_t v_start, uint16_t v_end, uint8_t h_start, uint8_t h_end);
 void Select_display_bus();
 void Select_control_bus();
